Movie Recommender System/Webapp/requirements.txt
streamlit==1.2.0
pandas==1.3.4
python-dotenv==0.19.2
requests==2.26.0